requirements-gpu.txt
absl-py==1.4.0
appdirs==1.4.4
blis==0.7.10
cachetools==5.3.1
catalogue==2.0.9
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
cmake==3.27.0
coloredlogs==15.0.1
confection==0.1.0
contourpy==1.1.0
cycler==0.11.0
cymem==2.0.7
docker-pycreds==0.4.0
einops==0.6.1
filelock==3.12.2
flash-attn==2.0.1
flatbuffers==23.5.26
fonttools==4.41.1
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.56.2
humanfriendly==10.0
idna==3.4
Jinja2==3.1.2
joblib==1.3.1
kiwisolver==1.4.4
langcodes==3.3.0
lit==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
murmurhash==1.0.9
networkx==3.1
ninja==1.11.1
numpy==1.25.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
onnx==1.14.0
onnxruntime-gpu==1.15.1
packaging==23.1
pandas==2.0.3
pathtools==0.1.2
pathy==0.10.2
Pillow==10.0.0
preshed==3.0.8
protobuf==4.23.4
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.12
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-triton==2.1.0+9e3e10c5ed
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
sentry-sdk==1.28.1
setproctitle==1.3.2
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
srsly==2.4.7
sympy==1.12
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
thinc==8.1.10
threadpoolctl==3.2.0
torch==2.1.0.dev20230727+cu121
torch-tb-profiler==0.4.1
torchaudio==2.1.0.dev20230727+cu121
torchvision==0.16.0.dev20230727+cu121
tqdm==4.65.0
triton==2.0.0
typer==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
uniplot==0.10.0
urllib3==2.0.4
wasabi==1.1.2
Werkzeug==2.3.6
